Se agregó: No se puede eliminar participación de jueces, anotaciones y sustituciones de un partido finalizado. Se aumentó la expiración del token de acceso a 4hs

+++ Backend Tesis/backendAdministradorCompetenciasFutbolisticas/src/main/java/backendAdministradorCompetenciasFutbolisticas/Service/AnotacionService.java
@@ -85,6 +85,9 @@ public class AnotacionService {
 
     public void eliminarAnotacion(Long id){
         Anotacion anotacion = getAnotacion(id);
+        if(anotacion.getPartido().getEstado().equals(NombreEstadoPartido.FINALIZADO)){
+            throw new BadRequestException("No se puede eliminar una anotación de un partido finalizado");
+        }
         anotacionRepository.deleteById(anotacion.getId());
     }
